requirements.txt
pyyaml
tqdm
scikit-learn
pillow
matplotlib
pandas
# Uncomment and set CUDA index for GPUs later:
# torch
# torchvision
# torchaudio
torch
torchvision
torchaudio
